chore: pr #2337 changes
